Create generate food and check if the snakes head its on top of the food
eatFood function checks if the snakes head its on top of the food, if it is then it will generate some more food. and prevent a nasty bug(loops) need to check if the bug still exists.
collisionDetection function need to clean this up a bit. and check if the snake collided wit the wall or it's self. if the snake collided with the wall, there's no point in checking if it collided with itself.
generateFood function is uploaded last time.